+++ AsteroidsFX/CommonBullet/src/main/java/dk/sdu/cbse/bulletsystem/BulletSPI.java
@@ -3,5 +3,5 @@ package dk.sdu.cbse.bulletsystem;
 import dk.sdu.cbse.common.Entity;
 
 public interface BulletSPI {
-    Entity CreateBullet(float[] position, float rotation);
+    Entity CreateBullet(double[] position, double rotation);
 }
